Reservation Container 분리
