+++ app-web/src/ReservationForm.tsx
@@ -1,95 +1,15 @@
-import axios from "axios";
-import {useState} from "react";
+import { useFrom } from 'react-hook-form';
 
 function ReservationForm() {
-    const [seatNumber, setSeatNumber] = useState("");
-    const [userName, setUserName] = useState("");
-    const [checkIn, setCheckIn] = useState("");
-    const [checkOut, setCheckOut] = useState("");
-
-    const onChangeSeatNumber = (e:any) => {
-        e.preventDefault();
-        setSeatNumber(e.target.value);
-    };
-    const onChangeUserName = (e:any) => {
-        e.preventDefault();
-        setUserName(e.target.value);
-    };
-    const onChangeCheckIn = (e:any) => {
-        e.preventDefault();
-        setCheckIn(e.target.value);
-    };
-    const onChangeCheckOut = (e:any) => {
-        e.preventDefault();
-        setCheckOut(e.target.value);
-    };
-
-    const handleSubmit = async(e:any) => {
-        e.preventDefault();
-
-        const body = {
-            name: userName,
-            checkIn: checkIn,
-            checkOut: checkOut
-        };
-
-        await axios.post('http://localhost:8080/seat/' + seatNumber,  {
-            body,
-        })
-            .then((response) => {
-                if(response.status === 201) {
-                    alert('예약 완료');
-                }
-            })
-            .catch((error) => {
-                console.log(error.response);
-
-                if(error.response.status === 400) {
-                    alert('이미 예약된 좌석입니다');
-                } else if(error.response.status === 404) {
-                    alert('존재하지 않는 좌석입니다');
-                }
-            });
-
-    }
-
+    const { register } = useFrom();
     return (
         <section>
-            <form onSubmit={handleSubmit}>
-                <input
-                    name="seatNumber"
-                    value={seatNumber}
-                    onChange={onChangeSeatNumber}
-                    type="text"
-                    placeholder="좌석 번호"
-                />
-
-                <input
-                    name="userName"
-                    value={userName}
-                    onChange={onChangeUserName}
-                    type="text"
-                    placeholder="이름"
-                />
-
-                <input
-                    name="checkIn"
-                    value={checkIn}
-                    onChange={onChangeCheckIn}
-                    type="text"
-                    placeholder="입실 시간(HH:MM)"
-                />
-
-                <input
-                    name="checkOut"
-                    value={checkOut}
-                    onChange={onChangeCheckOut}
-                    type="text"
-                    placeholder="퇴실 시간(HH:MM)"
-                />
-
-                <button type="submit">예약하기</button>
-            </form>
+          <form>
+            <input {...register("seatNumber")} placeholder="좌석 번호" />
+            <input {...register("name")} placeholder="이름" />
+            <input {...register("checkIn")} placeholder="예약 시작 시간(hh:mm)" />
+            <input {...register("name")} placeholder="이용 종료 시간(hh:mm)" />
+          </form>
         </section>
     );
 }
